Add beginning steps for network setup
Added the beginning steps for network setup, like creating the docker
network and backing up the resolv file. These steps don't have a way to
be undone just yet, so I still need to add that. In addition, testing
will be necessary as well, but that will likely come at the end of these
first couple of commits.