Add initial daemon stuff
In order for the proxy container running Traefik to work, it must be on
the same Docker network as the other containers it is acting as a
reverse-proxy for. There were two options: Create a default network and
then change the docker-compose for every project to use that network or
automatically have the proxy container join every Docker network that
gets created. I chose the second option, as this way the user doesn't
have to worry about polluting their docker-compose files as much and
cause daemon's are cool.

Note that the "falcon up" command doesn't currently create the proxy
container, so it's effectively useless. In addition, there's no
automated way to revert the changes it makes, so avoid using it for now.