Update daemon networking strategy and refactor
Previously, the daemon was only listening for network create and destroy
events. This meant that the proxy would never actually leave a Docker
network it was a part of, as a network can only ever be destroyed if it
has no containers on it. This new strategy has two parts to it:

1. When the daemon is first started, it performs an initial sync. This
   ensures that the proxy joins all of the networks it is supposed to.
   Otherwise, the daemon would connect the proxy to networks until a
   user started doing things in Docker land.
2. Listen to the network events connect and disconnect, and then simply
   syncs again, using the same methodology as specified in sync. While
   this is technically more expensive than determining whether to join
   or leave a network as connect's and disconnect's happen, it's also
   much more fault tolerant, as if one run of the daemon somehow misses
   an event or an error happens only in one run, then the next run of
   the daemon will catch it. It also requires much less code, and is
   exactly how the dinghy-http-proxy does its network autojoin magic.

This new methodology for the daemon should ensure that the proxy stays
in sync. What I both love and hate is a) the fact that I couldn't
improve on the dinghy-http-proxy way of doing things and had to copy
their code and b) that the dinghy-http-proxy already had a really good
way to solve this problem. Thanks FreedomBen!