Implement basic proxy starting logic
Implemented some very basic proxy pulling and starting logic. Now we
have a way to pull the falcon-proxy image and start it, with the correct
port mappings and volume mount to the Docker socket. A lot of it is
still hardcoded, but it'll do for now. I was able to successfully pull
the image, start it, create another docker container on the bridge
network with the appropriate Traefik labels, and then talk to it! That
was an exhilarating feeling.

Currently, the falcon up command is broken due to an issue with the
regex. I've even tried fixing it to `^\[main\]$`, but it still won't
recognize the main section in the NetworkManager file. Part of me still
wonders if doing the CLI in Node might've been a better idea. I liked
the idea of using an official Docker client, but it isn't any better
than say, dockerode. Neither of them has quite the documentation I would
like.