Update Networking
Updated the networking to now add the special loopback address sauce. In
addition, changed the directory structure so that we now can separate
out the logic needed to configure networking for macOS vs Linux. At this
point, I think that most of the networking for Linux should be done, but
I'm sure I'll discover bugs. The next step is to see if I can figure out
how to test some of these things and work on the macOS networking stuff.