Fix Docker and Shell bugs
Previously, my code wasn't checking to see if a container already
existed with the same name before trying to create and start a new
container. In addition, it wasn't able to remove the containers, as
ContainerRemove requires the container to be stopped before removing it.
Both of these bugs have now been fixed, which makes the CLI much
friendlier, as you can run `falcon up` as many times as you'd like
without any unnecessary errors being thrown.

In addition, the shell code wasn't properly executing commands as it was
only grabbing stdout, not stderr. This meant that we weren't catching
and recovering from expected errors. Now we grab both stdout and stderr,
so we can properly recover.