Pull out dependencies for docker package
Pulled out the implicit client that we were creating in all the docker
package functions. This will make for much much easier testing later on,
as almost everything depends on the docker package to actually do
things. Next up, time to pull up the dependencies in the
networking/darwin package!