Add first tests for docker package
Added in the first tests for the docker package. These should cover the
GetContainer and RemoveContainer methods. This required adding in the
ginkgo testing framework, the gomega assertion library, and the gomock
mocking library.

One thing I'm learning is that mocking in statically typed languages, or
at least a language like Go, can be rather painful. It requires a lot of
boilerplate and at least in Go, the IDE type inference suffers, likely
because of the current lack of generics.

I still need to add in a CI/CD workflow to get testing automated, so
that'll probably come soon.