Change to struct/interface for Docker interaction
I've changed the docker package to a more object oriented style for
interacting with the Docker API. This required only minor changes to the
function/method signatures and will make it significantly easier to mock
interacting with the Docker API for future tests, such as for testing the
dnsmasq and proxy packages. While first defining the interface and then
making the actual struct that implements this is rather frustrating,
this is the price you pay for statically typed, compiled languages.

Hopefully, I can use this similar style of switching to structs to make
it testing the files and networking packages easier, but we'll see.