Finish dnsmasq tests, continue work on docker test
The dnsmasq package tests are finished, seeing as they were extremely
easy to write. However, I'm still working on setting up all of the
necessary stuff to test the StartContainer method for the DockerClient
method. It's a rather complex method, with a lot of cases to cover. More
to come!