Add networking tests and remove Linux support
Added in tests for the darwin networking code, while also changing it to
be more testable. Note that I have also removed support for Linux, at
least temporarily. This is due to the difficulties I'll likely face when
rewriting the Linux code to be more testable as well as the fact that
Linux isn't used much for development at my work, so it just doesn't
seem worth it currently.