Switch to directory bind-mount for config
Apparently, docker sometimes has issues with fsnotify if you bind mount
a file directly into a container. This meant that Traefik wasn't picking
up on our changes to the dynamic config, so you couldn't dynamically add
a certificate. Hopefully, by switching to a directory bind-mount, we can
fix this issue. This change is mostly renaming constants to use the new
path.